move auth tests to respective file
